Fix some new lint errors
