+++ Sources/AuthFoundation/JWT/JWKS.swift
@@ -37,8 +37,8 @@ public struct JWKS: Codable, Equatable, Collection {
         keys[index]
     }
     
-    public func index(after i: Index) -> Index {
-        keys.index(after: i)
+    public func index(after index: Index) -> Index {
+        keys.index(after: index)
     }
 
     public static func == (lhs: JWKS, rhs: JWKS) -> Bool {
