+++ Sources/AuthFoundation/Version.swift
@@ -12,4 +12,6 @@
 
 import Foundation
 
+// swiftlint:disable identifier_name
 public let Version = SDKVersion(sdk: "okta-authfoundation-swift", version: "1.3.1")
+// swiftlint:enable identifier_name
