+++ Sources/OktaOAuth2/Version.swift
@@ -12,4 +12,6 @@
 
 @_exported import AuthFoundation
 
+// swiftlint:disable identifier_name
 public let Version = SDKVersion(sdk: "okta-oauth2-swift", version: "1.3.1")
+// swiftlint:enable identifier_name
