+++ Sources/WebAuthenticationUI/Version.swift
@@ -13,4 +13,6 @@
 import Foundation
 import AuthFoundation
 
+// swiftlint:disable identifier_name
 public let Version = SDKVersion(sdk: "okta-webauthenticationui-swift", version: "1.3.1")
+// swiftlint:enable identifier_name
